backend/requirements.txt
# Web Framework
Flask
Flask-Cors
Flask-JWT-Extended
gunicorn

# Database
psycopg2-binary

# Environment and Security
python-dotenv
bcrypt

# HTTP and API
requests

# AI and Machine Learning
transformers

# Media Processing
youtube-transcript-api
Pillow

# PDF Generation and Charts
reportlab
matplotlib
numpy

# Additional Dependencies
Werkzeug
